F move persistence mocks from test to production package
