+++ approvaltests-util/src/main/java/com/spun/util/persistence/test/MockDeletor.java
@@ -1,18 +1,14 @@
-package com.spun.util.persistence.tests;
+package com.spun.util.persistence.test;
 
 import com.spun.util.persistence.Deletor;
 
 public class MockDeletor<T> implements Deletor<T>
 {
-
   private T deleted;
-
   public void delete(T delete)
   {
     this.deleted = delete;
-    
   }
-
   public T getDeleted()
   {
     return deleted;
