+++ approvaltests-util/src/main/java/com/spun/util/persistence/test/MockLoader.java
@@ -1,4 +1,4 @@
-package com.spun.util.persistence.tests;
+package com.spun.util.persistence.test;
 
 import com.spun.util.persistence.Loader;
