+++ approvaltests-util/src/main/java/com/spun/util/persistence/test/MockSaver.java
@@ -1,9 +1,8 @@
-package com.spun.util.persistence.tests;
-
-import java.util.ArrayList;
+package com.spun.util.persistence.test;
 
 import com.spun.util.ArrayUtils;
 import com.spun.util.persistence.Saver;
+import java.util.ArrayList;
 
 public class MockSaver<T> implements Saver<T>
 {
